# Raise interrupt for used descriptors on tx queue and use edged triggered epoll.